allow any custom payload publisher
