+++ .c8rc.json
@@ -4,7 +4,8 @@
   "exclude": [
     "src/**/index.ts",
     "src/jsutils/ObjMap.ts",
-    "src/jsutils/PromiseOrValue.ts"
+    "src/jsutils/PromiseOrValue.ts",
+    "src/transform/PayloadPublisher.ts"
   ],
   "clean": true,
   "report-dir": "reports/coverage",
