+++ src/transform/__tests__/cancellation-test.ts
@@ -11,7 +11,7 @@ import { invariant } from '../../jsutils/invariant.js';
 import type {
   LegacyInitialIncrementalExecutionResult,
   LegacySubsequentIncrementalExecutionResult,
-} from '../PayloadPublisher.js';
+} from '../legacyExecuteIncrementally.js';
 
 import { execute } from './execute.js';
