+++ src/transform/transformResult.ts
@@ -1,35 +1,59 @@
 import type {
+  ExecutionArgs,
   ExecutionResult,
   ExperimentalIncrementalExecutionResults,
   GraphQLError,
 } from 'graphql';
+import { experimentalExecuteQueryOrMutationOrSubscriptionEvent } from 'graphql';
+// eslint-disable-next-line n/no-missing-import
+import { validateExecutionArgs } from 'graphql/execution/execute.js';
 import type {
   PendingResult,
   SubsequentIncrementalExecutionResult,
   // eslint-disable-next-line n/no-missing-import
 } from 'graphql/execution/types.js';
 
+import { isPromise } from '../jsutils/isPromise.js';
 import type { ObjMap } from '../jsutils/ObjMap.js';
+import type { PromiseOrValue } from '../jsutils/PromiseOrValue.js';
 
 import type { TransformationContext } from './buildTransformationContext.js';
+import { buildTransformationContext } from './buildTransformationContext.js';
 import { completeInitialResult } from './completeValue.js';
 import { embedErrors } from './embedErrors.js';
 import { IncrementalPublisher } from './IncrementalPublisher.js';
-import type {
-  LegacyExperimentalIncrementalExecutionResults,
-  LegacySubsequentIncrementalExecutionResult,
-  PayloadPublisher,
-} from './PayloadPublisher.js';
+import type { PayloadPublisher } from './PayloadPublisher.js';
 import type { IncrementalDataRecord } from './types.js';
 
-export function transformResult(
+export function transformResult<TSubsequent, TIncremental>(
+  args: ExecutionArgs,
+  payloadPublisher: PayloadPublisher<TSubsequent, TIncremental>,
+  prefix = '__transformResult__',
+): PromiseOrValue<ExecutionResult | TIncremental> {
+  const originalArgs = validateExecutionArgs(args);
+
+  if (!('schema' in originalArgs)) {
+    return { errors: originalArgs };
+  }
+
+  const context = buildTransformationContext(originalArgs, prefix);
+
+  const originalResult = experimentalExecuteQueryOrMutationOrSubscriptionEvent(
+    context.transformedArgs,
+  );
+
+  return isPromise(originalResult)
+    ? originalResult.then((resolved) =>
+        transformOriginalResult(context, resolved, payloadPublisher),
+      )
+    : transformOriginalResult(context, originalResult, payloadPublisher);
+}
+
+function transformOriginalResult<TSubsequent, TIncremental>(
   context: TransformationContext,
   result: ExecutionResult | ExperimentalIncrementalExecutionResults,
-  payloadPublisher: PayloadPublisher<
-    LegacySubsequentIncrementalExecutionResult,
-    LegacyExperimentalIncrementalExecutionResults
-  >,
-): ExecutionResult | LegacyExperimentalIncrementalExecutionResults {
+  payloadPublisher: PayloadPublisher<TSubsequent, TIncremental>,
+): ExecutionResult | TIncremental {
   if ('initialResult' in result) {
     const { initialResult, subsequentResults } = result;
     const { data, errors, pending } = initialResult;
@@ -62,20 +86,17 @@ export function transformResult(
   return buildIncrementalResponse(originalData, completed, payloadPublisher);
 }
 
-function buildIncrementalResponse(
+function buildIncrementalResponse<TSubsequent, TIncremental>(
   originalData: ObjMap<unknown>,
   initialResult: {
     data: ObjMap<unknown>;
     errors: ReadonlyArray<GraphQLError>;
     incrementalDataRecords: ReadonlyArray<IncrementalDataRecord>;
   },
-  payloadPublisher: PayloadPublisher<
-    LegacySubsequentIncrementalExecutionResult,
-    LegacyExperimentalIncrementalExecutionResults
-  >,
+  payloadPublisher: PayloadPublisher<TSubsequent, TIncremental>,
   pending?: ReadonlyArray<PendingResult>,
   subsequentResults?: AsyncGenerator<SubsequentIncrementalExecutionResult>,
-): LegacyExperimentalIncrementalExecutionResults {
+): TIncremental {
   const { data, errors, incrementalDataRecords } = initialResult;
 
   const incrementalPublisher = new IncrementalPublisher(
